+++ Data/Purges/purge_tbc.lua
@@ -1,7 +1,9 @@
 -- Auto-generated data file. Do not edit manually.
--- Version: a9bf4107
+-- Version: faf90472
 if WOW_PROJECT_ID ~= WOW_PROJECT_BURNING_CRUSADE_CLASSIC then return end
 
 local _dd = DeathlogData.Internal ---@class _dd
 if not _dd then return end
-_dd.PRECOMPUTED_PURGES = {}
+_dd.PRECOMPUTED_PURGES = {
+	Thunderstrike = {},
+}