+++ Libs/DeathNotificationLib/Data/DeathNotificationLibData~Finalizer.lua
@@ -1,0 +1,1 @@
+if DeathNotificationLibData.Internal then DeathNotificationLibData.Internal = nil end