+++ Libs/DeathNotificationLib/DeathNotificationLib~Broadcast.lua
@@ -314,6 +314,7 @@ function _dnl.reportPartyRaidDeath(unit)
 	local death_source, extra_data = _dnl.resolveDeathSource(unitState.last_attack_source_name, unitState.last_attack_source_guid)
 
 	local instance_id, map, position = _dnl.guessLocationForUnit(unit)
+	local pos_str = position and string.format("%.4f,%.4f", position.x, position.y) or nil
 
 	local last_words = unitState.recent_msg
 
@@ -326,7 +327,7 @@ function _dnl.reportPartyRaidDeath(unit)
 		level,
 		instance_id,
 		map,
-		position,
+		pos_str,
 		GetServerTime(),
 		nil,
 		last_words,