+++ deathlog.lua
@@ -412,24 +412,40 @@ local function handleEvent(self, event, ...)
 		})
 		DeathNotificationLib.HookOnNewAddonEntry("Deathlog", newEntry)
 		if DeathlogDataCopy.PRECOMPUTED_GENERAL_STATS and DeathlogDataCopy.PRECOMPUTED_LOG_NORMAL_PARAMS and DeathlogDataCopy.PRECOMPUTED_KAPLAN_MEIER and DeathlogDataCopy.PRECOMPUTED_MOST_DEADLY_BY_ZONE and DeathlogDataCopy.PRECOMPUTED_PURGES then
-			for k in pairs(deathlog_precomputed) do
-				deathlog_precomputed[k] = nil
-			end
+			deathlog_precomputed.PRECOMPUTED_GENERAL_STATS = nil
+			deathlog_precomputed.PRECOMPUTED_LOG_NORMAL_PARAMS = nil
+			deathlog_precomputed.PRECOMPUTED_KAPLAN_MEIER = nil
+			deathlog_precomputed.PRECOMPUTED_MOST_DEADLY_BY_ZONE = nil
+			deathlog_precomputed.PRECOMPUTED_PURGES = nil
 		else
-			local computed = false
-			if not (version == deathlog_precomputed.VERSION and deathlog_precomputed.PRECOMPUTED_GENERAL_STATS and deathlog_precomputed.PRECOMPUTED_LOG_NORMAL_PARAMS and deathlog_precomputed.PRECOMPUTED_KAPLAN_MEIER and deathlog_precomputed.PRECOMPUTED_MOST_DEADLY_BY_ZONE and deathlog_precomputed.PRECOMPUTED_PURGES) then
-				deathlog_precomputed.VERSION = version
+			if not (version == deathlog_precomputed.DD_VERSION and deathlog_precomputed.PRECOMPUTED_GENERAL_STATS and deathlog_precomputed.PRECOMPUTED_LOG_NORMAL_PARAMS and deathlog_precomputed.PRECOMPUTED_KAPLAN_MEIER and deathlog_precomputed.PRECOMPUTED_MOST_DEADLY_BY_ZONE and deathlog_precomputed.PRECOMPUTED_PURGES) then
+				deathlog_precomputed.DD_VERSION = version
 				deathlog_precomputed.PRECOMPUTED_GENERAL_STATS = Deathlog_calculate_statistics(deathlog_data)
 				deathlog_precomputed.PRECOMPUTED_LOG_NORMAL_PARAMS = Deathlog_calculateLogNormalParameters(deathlog_data)
 				deathlog_precomputed.PRECOMPUTED_KAPLAN_MEIER = Deathlog_calculateKaplanMeier(deathlog_data)
 				deathlog_precomputed.PRECOMPUTED_MOST_DEADLY_BY_ZONE = Deathlog_calculateMostDeadlyByZone(deathlog_data)
 				deathlog_precomputed.PRECOMPUTED_PURGES = Deathlog_calculatePurges(deathlog_data)
-				computed = true
 			end
 
-			for k, v in pairs(deathlog_precomputed) do
-				DeathlogDataCopy[k] = v
+			DeathlogDataCopy.PRECOMPUTED_GENERAL_STATS = deathlog_precomputed.PRECOMPUTED_GENERAL_STATS
+			DeathlogDataCopy.PRECOMPUTED_LOG_NORMAL_PARAMS = deathlog_precomputed.PRECOMPUTED_LOG_NORMAL_PARAMS
+			DeathlogDataCopy.PRECOMPUTED_KAPLAN_MEIER = deathlog_precomputed.PRECOMPUTED_KAPLAN_MEIER
+			DeathlogDataCopy.PRECOMPUTED_MOST_DEADLY_BY_ZONE = deathlog_precomputed.PRECOMPUTED_MOST_DEADLY_BY_ZONE
+			DeathlogDataCopy.PRECOMPUTED_PURGES = deathlog_precomputed.PRECOMPUTED_PURGES
+		end
+		if DeathNotificationLibDataCopy.HEATMAP_INTENSITY and DeathNotificationLibDataCopy.HEATMAP_CREATURE_SUBSET then
+			deathlog_precomputed.HEATMAP_INTENSITY = nil
+			deathlog_precomputed.HEATMAP_CREATURE_SUBSET = nil
+		else
+			if not (version == deathlog_precomputed.DNLD_VERSION and deathlog_precomputed.HEATMAP_INTENSITY and deathlog_precomputed.HEATMAP_CREATURE_SUBSET) then
+				local skull_locs = Deathlog_calculateSkullLocs(deathlog_data)
+				deathlog_precomputed.DNLD_VERSION = version
+				deathlog_precomputed.HEATMAP_INTENSITY = Deathlog_calculateHeatmapIntensity(skull_locs)
+				deathlog_precomputed.HEATMAP_CREATURE_SUBSET = Deathlog_calculateHeatmapCreatureSubset(skull_locs)
 			end
+
+			DeathNotificationLibDataCopy.HEATMAP_INTENSITY = deathlog_precomputed.HEATMAP_INTENSITY
+			DeathNotificationLibDataCopy.HEATMAP_CREATURE_SUBSET = deathlog_precomputed.HEATMAP_CREATURE_SUBSET
 		end
 		loadWidgets()